Fix Q-Block clippy conversion and Q-Block2 option order
trailing_ones() is already u32. Size2 (28) must precede Q-Block2 (31)
on the encode path.

+++ src/storage/block.rs
@@ -658,7 +658,7 @@ impl BlockTransfer {
 
         let contig = q
             .base
-            .checked_add(u32::from(q.mask.trailing_ones()))
+            .checked_add(q.mask.trailing_ones())
             .ok_or(BlockTransferError::Overflow)?;
         if let Some(final_num) = q.final_num {
             if contig > final_num {
